Refactor to NDK 26 and use ANDROID_WEAK_API_DEFS
