+++ build.gradle
@@ -1,6 +1,6 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 plugins {
-    id 'com.android.application' version '8.1.1' apply false
+    id 'com.android.application' version '8.3.1' apply false
     id 'org.jetbrains.kotlin.android' version '1.9.10' apply false
 }
 ext {
