+++ libraw/build.gradle
@@ -7,7 +7,8 @@ android {
     //ndkVersion "22.1.7171670"
     //Required for __fp16 support
     //ndkVersion "23.0.7599858"
-    ndkVersion "23.1.7779620"
+    //Required for __ANDROID_UNAVAILABLE_SYMBOLS_ARE_WEAK__
+    ndkVersion "26.3.11579264"
     defaultConfig {
         //Has to be 24 for 64-bit functions
         minSdk 24
@@ -18,10 +19,10 @@ android {
         externalNativeBuild {
             cmake {
                 arguments "-DLIBRAW_PATH=${projectDir}/src/main/cpp/LibRaw/",
-                        //Hack to allow building LibRaw26
-                        "-DANDROID_PLATFORM=26",
-                        // You can manually include openMP by renameing jniLibsBak to jniLibs and commenting the entry below.
-                        // I noticed not performance increase, so I disabled it to reduce APK size
+                        "-DANDROID_WEAK_API_DEFS=ON",
+                        "-Werror=unguarded-availability",
+                        // You can manually include openMP by renaming jniLibsBak to jniLibs and commenting the entry below.
+                        // I noticed no performance increase, so I disabled it to reduce APK size
                         // Also, the openMP build is broken, so it seems like it's fate.
                         "-DENABLE_OPENMP=OFF",
                         //Disable Jasper (JPEG 2000)
@@ -76,7 +77,7 @@ publishing {
 }
 
 dependencies {
-    implementation 'androidx.annotation:annotation:1.7.0'
+    implementation 'androidx.annotation:annotation:1.7.1'
     testImplementation 'junit:junit:4.+'
     androidTestImplementation 'androidx.test.ext:junit:1.1.5'
 }
